+++ src/main/java/org/smoothbuild/parse/FindSemanticErrors.java
@@ -1,5 +1,6 @@
 package org.smoothbuild.parse;
 
+import static java.util.Collections.sort;
 import static java.util.stream.Collectors.toMap;
 import static java.util.stream.Collectors.toSet;
 import static org.smoothbuild.util.Lists.map;
@@ -15,28 +16,39 @@ import java.util.Set;
 import org.smoothbuild.lang.function.Functions;
 import org.smoothbuild.lang.function.base.Function;
 import org.smoothbuild.lang.message.Location;
+import org.smoothbuild.lang.runtime.SRuntime;
+import org.smoothbuild.lang.type.RuntimeTypes;
 import org.smoothbuild.parse.ast.ArgNode;
+import org.smoothbuild.parse.ast.ArrayTypeNode;
 import org.smoothbuild.parse.ast.Ast;
 import org.smoothbuild.parse.ast.CallNode;
+import org.smoothbuild.parse.ast.FieldNode;
 import org.smoothbuild.parse.ast.FuncNode;
+import org.smoothbuild.parse.ast.Named;
 import org.smoothbuild.parse.ast.NamedNode;
 import org.smoothbuild.parse.ast.ParamNode;
 import org.smoothbuild.parse.ast.RefNode;
 import org.smoothbuild.parse.ast.StringNode;
+import org.smoothbuild.parse.ast.StructNode;
+import org.smoothbuild.parse.ast.TypeNode;
 import org.smoothbuild.util.UnescapingFailedException;
 
 import com.google.common.collect.ImmutableSet;
 
 public class FindSemanticErrors {
-  public static List<ParseError> findSemanticErrors(Functions functions, Ast ast) {
+  public static List<ParseError> findSemanticErrors(SRuntime runtime, Ast ast) {
     List<ParseError> errors = new ArrayList<>();
+    Functions functions = runtime.functions();
     unescapeStrings(errors, ast);
     parametersReferenceWithParentheses(errors, ast);
-    undefinedElements(errors, functions, ast);
-    duplicateFunctions(errors, functions, ast);
+    undefinedReferences(errors, functions, ast);
+    undefinedTypes(errors, runtime.types(), ast);
+    duplicateNames(errors, functions, ast);
+    duplicateFieldNames(errors, ast);
     duplicateParamNames(errors, ast);
     duplicateArgNames(errors, ast);
     unknownArgNames(errors, functions, ast);
+    firstFieldWithForbiddenType(errors, ast);
     return errors;
   }
 
@@ -67,7 +79,7 @@ public class FindSemanticErrors {
     }.visitAst(ast);
   }
 
-  private static void undefinedElements(List<ParseError> errors, Functions functions, Ast ast) {
+  private static void undefinedReferences(List<ParseError> errors, Functions functions, Ast ast) {
     Set<String> all = ImmutableSet.<String> builder()
         .addAll(functions.names())
         .addAll(map(ast.funcs(), f -> f.name()))
@@ -83,21 +95,64 @@ public class FindSemanticErrors {
     }.visitAst(ast);
   }
 
-  private static void duplicateFunctions(List<ParseError> errors, Functions functions, Ast ast) {
+  private static void undefinedTypes(List<ParseError> errors, RuntimeTypes types, Ast ast) {
+    Set<String> all = ImmutableSet.<String> builder()
+        .addAll(types.names())
+        .addAll(map(ast.structs(), s -> s.name()))
+        .build();
+    new AstVisitor() {
+      @Override
+      public void visitField(FieldNode field) {
+        super.visitField(field);
+        assertTypeIsDefined(field.type());
+      }
+
+      private void assertTypeIsDefined(TypeNode type) {
+        if (type instanceof ArrayTypeNode) {
+          assertTypeIsDefined(((ArrayTypeNode) type).elementType());
+        } else if (!all.contains(type.name())) {
+          errors.add(new ParseError(type.location(), "'" + type.name() + "' is undefined."));
+        }
+      }
+    }.visitAst(ast);
+  }
+
+  private static void duplicateNames(List<ParseError> errors, Functions functions, Ast ast) {
     Map<String, Location> defined = functions
         .functions()
         .stream()
         .collect(toMap(Function::name, Function::location));
+    List<Named> nodes = new ArrayList<>();
+    nodes.addAll(ast.structs());
+    nodes.addAll(ast.funcs());
+    sort(nodes, (node1, node2) -> {
+      int l1 = node1.location().line();
+      int l2 = node2.location().line();
+      if (l1 < l2) {
+        return -1;
+      } else if (l1 == l2) {
+        return 0;
+      } else {
+        return 1;
+      }
+    });
+    for (Named node : nodes) {
+      String name = node.name();
+      if (defined.containsKey(name)) {
+        errors.add(new ParseError(node.location(),
+            "'" + name + "' is already defined at " + defined.get(name) + "."));
+      } else {
+        defined.put(name, node.location());
+      }
+    }
+  }
+
+  private static void duplicateFieldNames(List<ParseError> errors, Ast ast) {
     new AstVisitor() {
       @Override
-      public void visitFunc(FuncNode func) {
-        super.visitFunc(func);
-        String name = func.name();
-        if (defined.containsKey(name)) {
-          errors.add(new ParseError(func, "Function '" + name
-              + "' is already defined at " + defined.get(name) + "."));
-        }
-        defined.put(name, func.location());
+      public void visitFields(List<FieldNode> fields) {
+        super.visitFields(fields);
+        findDuplicateNames(errors, fields);
       }
     }.visitAst(ast);
   }
@@ -173,6 +228,26 @@ public class FindSemanticErrors {
               .collect(toSet());
         }
         return null;
+      }
+    }.visitAst(ast);
+  }
+
+  private static void firstFieldWithForbiddenType(List<ParseError> errors, Ast ast) {
+    new AstVisitor() {
+      @Override
+      public void visitStruct(StructNode struct) {
+        super.visitStruct(struct);
+        List<FieldNode> fields = struct.fields();
+        if (!fields.isEmpty()) {
+          FieldNode field = fields.get(0);
+          TypeNode type = field.type();
+          if (type.name().equals("Nothing")) {
+            errors.add(new ParseError(field, "First field of struct cannot have 'Nothing' type."));
+          }
+          if (type instanceof ArrayTypeNode) {
+            errors.add(new ParseError(field, "First field of struct cannot have array type."));
+          }
+        }
       }
     }.visitAst(ast);
   }