+++ src/main/java/org/smoothbuild/parse/ast/TypeNode.java
@@ -2,15 +2,8 @@ package org.smoothbuild.parse.ast;
 
 import org.smoothbuild.lang.message.Location;
 
-public class TypeNode extends Node {
-  private final String name;
-
+public class TypeNode extends NamedNode {
   public TypeNode(String name, Location location) {
-    super(location);
-    this.name = name;
-  }
-
-  public String name() {
-    return name;
+    super(name, location);
   }
 }