+++ src/main/test/org/smoothbuild/parse/ast/AstTest.java
@@ -18,14 +18,14 @@ public class AstTest {
   @Test
   public void contains_function_passed_to_constructor() throws Exception {
     given(func = function("name"));
-    given(ast = new Ast(list(func)));
+    given(ast = new Ast(list(), list(func)));
     when(() -> ast.containsFunc("name"));
     thenReturned(true);
   }
 
   @Test
   public void does_not_contain_function_not_passed_to_constructor() throws Exception {
-    given(ast = new Ast(list()));
+    given(ast = new Ast(list(), list()));
     when(() -> ast.containsFunc("name"));
     thenReturned(false);
   }
@@ -33,14 +33,14 @@ public class AstTest {
   @Test
   public void function_passed_to_constructor_can_be_retrieved() throws Exception {
     given(func = function("name"));
-    given(ast = new Ast(list(func)));
+    given(ast = new Ast(list(), list(func)));
     when(() -> ast.func("name"));
     thenReturned(func);
   }
 
   @Test
   public void retrieving_function_not_contained_causes_exception() throws Exception {
-    given(ast = new Ast(list()));
+    given(ast = new Ast(list(), list()));
     when(() -> ast.func("name"));
     thenThrown(IllegalStateException.class);
   }